Fixed port test value type

+++ xabberTests/AccountTests.m
@@ -87,7 +87,7 @@
     acc.accountJID = @"account";
 
     XCTAssertTrue(acc.autoLogin);
-    XCTAssertEqual(acc.port, (UInt16)5222);
+    XCTAssertEqual(acc.port, 5222u);
     XCTAssertEqual(acc.status, XBAccountStatusAvailable);
     XCTAssertTrue(acc.isNew);
     XCTAssertFalse(acc.isDeleted);
